change mpi->divideWorkIter to return an std::vector of size_t values -- when very large calculations are done, int will cause overflow

+++ src/mpi/mpiController.cpp
@@ -96,21 +96,21 @@ void MPIcontroller::barrier() const {
 }
 
 // Labor division functions -----------------------------------------
-std::vector<int> MPIcontroller::divideWork(size_t numTasks) {
+std::vector<size_t> MPIcontroller::divideWork(size_t numTasks) {
   // return a vector of the start and stop points for task division
-  std::vector<int> divs(2);
+  std::vector<size_t> divs(2);
   divs[0] = (numTasks * rank) / size;
   divs[1] = (numTasks * (rank + 1)) / size;
   return divs;
 }
 
-std::vector<int> MPIcontroller::divideWorkIter(size_t numTasks) {
+std::vector<size_t> MPIcontroller::divideWorkIter(size_t numTasks) {
   // return a vector of indices for tasks to be completed by thsi process
-  int start = (numTasks * rank) / size;
-  int stop = (numTasks * (rank + 1)) / size;
+  size_t start = (numTasks * rank) / size;
+  size_t stop = (numTasks * (rank + 1)) / size;
   size_t localSize = stop - start;
-  std::vector<int> divs(localSize);
-  for (int i = start; i < stop; i++) {
+  std::vector<size_t> divs(localSize);
+  for (size_t i = start; i < stop; i++) {
     divs[i - start] = i;
   }
   return divs;
